change the base class
